+++ core/__init__.py
@@ -2,10 +2,11 @@ from .certification import Certifier, CertificationReport, calculate_tier
 from .strategy import StrategyAdapter, StrategyMetadata
 from .testing import CriteriaPolicy, TestCatalog, TestDefinition, default_catalog
 from .integrations import validate_dal_handoff
+from .metric import MetricAdapter, MetricMetadata
 
 __version__ = "0.1.0.dev1"
 
 __all__ = [
     "Certifier", "CertificationReport", "calculate_tier", "StrategyAdapter", "StrategyMetadata",
-    "CriteriaPolicy", "TestCatalog", "TestDefinition", "default_catalog", "validate_dal_handoff", "__version__",
+    "CriteriaPolicy", "MetricAdapter", "MetricMetadata", "TestCatalog", "TestDefinition", "default_catalog", "validate_dal_handoff", "__version__",
 ]